+++ packages/nodejs/test/accountTransaction.test.ts
@@ -7,13 +7,18 @@ import {
     AttributeKey,
     IdentityInput,
     IndexedCredentialDeploymentInfo,
+    Schedule,
     RegisterDataPayload,
     SimpleTransferPayload,
     SimpleTransferWithMemoPayload,
+    TransferWithSchedulePayload,
     UpdateCredentialsPayload,
     VerifyKey,
     ConfigureDelegationPayload,
     DelegationTargetType,
+    createEncryptedTransferPayload,
+    buildSignedCredentialForExistingAccount,
+    createUnsignedCredentialForExistingAccount,
 } from '@concordium/common-sdk';
 import * as ed from 'noble-ed25519';
 import {
@@ -25,12 +30,22 @@ import { AccountAddress } from '@concordium/common-sdk';
 import { GtuAmount } from '@concordium/common-sdk';
 import { TransactionExpiry } from '@concordium/common-sdk';
 import { DataBlob } from '@concordium/common-sdk';
-import {
-    buildSignedCredentialForExistingAccount,
-    createUnsignedCredentialForExistingAccount,
-} from '@concordium/common-sdk';
 
 const client = getNodeClient();
+
+async function getAccountHeader(
+    sender: AccountAddress
+): Promise<AccountTransactionHeader> {
+    const nextAccountNonce = await client.getNextAccountNonce(sender);
+    if (!nextAccountNonce) {
+        throw new Error('Nonce not found!');
+    }
+    return {
+        expiry: new TransactionExpiry(new Date(Date.now() + 3600000)),
+        nonce: nextAccountNonce.nonce,
+        sender,
+    };
+}
 
 test('send simple transfer signed with wrong private key is accepted', async () => {
     const senderAccountAddress =
@@ -42,17 +57,9 @@ test('send simple transfer signed with wrong private key is accepted', async () 
         ),
     };
 
-    const nextAccountNonce = await client.getNextAccountNonce(
+    const header = await getAccountHeader(
         new AccountAddress(senderAccountAddress)
     );
-    if (!nextAccountNonce) {
-        throw new Error('Nonce not found!');
-    }
-    const header: AccountTransactionHeader = {
-        expiry: new TransactionExpiry(new Date(Date.now() + 3600000)),
-        nonce: nextAccountNonce.nonce,
-        sender: new AccountAddress(senderAccountAddress),
-    };
 
     const simpleTransferAccountTransaction: AccountTransaction = {
         header: header,
@@ -93,17 +100,9 @@ test('send simple transfer with memo signed with wrong private key is accepted',
         memo: new DataBlob(Buffer.from('6B68656C6C6F20776F726C64', 'hex')),
     };
 
-    const nextAccountNonce = await client.getNextAccountNonce(
+    const header = await getAccountHeader(
         new AccountAddress(senderAccountAddress)
     );
-    if (!nextAccountNonce) {
-        throw new Error('Nonce not found!');
-    }
-    const header: AccountTransactionHeader = {
-        expiry: new TransactionExpiry(new Date(Date.now() + 3600000)),
-        nonce: nextAccountNonce.nonce,
-        sender: new AccountAddress(senderAccountAddress),
-    };
 
     const simpleTransferAccountTransaction: AccountTransaction = {
         header: header,
@@ -202,16 +201,9 @@ test('update credential is accepted', async () => {
         ],
     };
 
-    const nextAccountNonce = await client.getNextAccountNonce(address);
-    if (!nextAccountNonce) {
-        throw new Error('Nonce not found');
-    }
-
-    const header: AccountTransactionHeader = {
-        expiry: new TransactionExpiry(new Date(Date.now() + 3600000)),
-        nonce: nextAccountNonce.nonce,
-        sender: address,
-    };
+    const header = await getAccountHeader(
+        new AccountAddress(senderAccountAddress)
+    );
 
     const updateCredentialsAccountTransaction: AccountTransaction = {
         header: header,
@@ -238,6 +230,105 @@ test('update credential is accepted', async () => {
         updateCredentialsAccountTransaction,
         signatures
     );
+    expect(result).toBeTruthy();
+});
+
+test('send transfer with schedule signed with wrong private key is accepted', async () => {
+    const senderAccountAddress =
+        '4ZJBYQbVp3zVZyjCXfZAAYBVkJMyVj8UKUNj9ox5YqTCBdBq2M';
+
+    const header = await getAccountHeader(
+        new AccountAddress(senderAccountAddress)
+    );
+
+    const schedule: Schedule = [
+        {
+            timestamp: new Date(Date.now() + 36000000),
+            amount: new GtuAmount(50n),
+        },
+        {
+            timestamp: new Date(Date.now() + 36500000),
+            amount: new GtuAmount(25n),
+        },
+    ];
+
+    const scheduledTransfer: TransferWithSchedulePayload = {
+        toAddress: new AccountAddress(
+            '4hXCdgNTxgM7LNm8nFJEfjDhEcyjjqQnPSRyBS9QgmHKQVxKRf'
+        ),
+        schedule: schedule,
+    };
+
+    const scheduledTransferAccountTransaction: AccountTransaction = {
+        header: header,
+        payload: scheduledTransfer,
+        type: AccountTransactionType.TransferWithSchedule,
+    };
+
+    const wrongPrivateKey =
+        'ce432f6cca0d47caec1f45739331dc354b6d749fdb8ab7c2b7f6cb24db39ca0c';
+
+    const hashToSign = getAccountTransactionSignDigest(
+        scheduledTransferAccountTransaction
+    );
+
+    const signature = Buffer.from(
+        await ed.sign(hashToSign, wrongPrivateKey)
+    ).toString('hex');
+    const signatures: AccountTransactionSignature = {
+        0: {
+            0: signature,
+        },
+    };
+
+    const result = await client.sendAccountTransaction(
+        scheduledTransferAccountTransaction,
+        signatures
+    );
+    expect(result).toBeTruthy();
+});
+
+test('send shielded transfer signed with wrong private key is accepted', async () => {
+    const sender = new AccountAddress(
+        '4EdBeGmpnQZWxaiig7FGEhWwmJurYmYsPWXo6owMDxA7ZtJMMH'
+    );
+    const receiver = new AccountAddress(
+        '4hXCdgNTxgM7LNm8nFJEfjDhEcyjjqQnPSRyBS9QgmHKQVxKRf'
+    );
+    // This is the actual decryptionKey of the sender. Otherwise creating the encryptedTransferData would never terminate.
+    const decryptionKey =
+        'b14cbfe44a02c6b1f78711176d5f437295367aa4f2a8c2551ee10d25a03adc69d61a332a058971919dad7312e1fc94c54f10b8b7388dbeefe1e98ac22e6041c2fb92e1562a59e04a03fa0ebc0a889e72';
+
+    const payload = await createEncryptedTransferPayload(
+        sender,
+        receiver,
+        new GtuAmount(50n),
+        decryptionKey,
+        client
+    );
+
+    const header = await getAccountHeader(sender);
+
+    const transaction: AccountTransaction = {
+        header,
+        payload,
+        type: AccountTransactionType.EncryptedTransfer,
+    };
+
+    const wrongPrivateKey =
+        'ce432f6cca0d47caec1f45739331dc354b6d749fdb8ab7c2b7f6cb24db39ca0c';
+
+    const hashToSign = getAccountTransactionSignDigest(transaction);
+    const signature = Buffer.from(
+        await ed.sign(hashToSign, wrongPrivateKey)
+    ).toString('hex');
+    const signatures: AccountTransactionSignature = {
+        0: {
+            0: signature,
+        },
+    };
+
+    const result = await client.sendAccountTransaction(transaction, signatures);
     expect(result).toBeTruthy();
 });
 
@@ -270,6 +361,7 @@ test('send registerData signed with wrong private key is accepted', async () => 
         'ce432f6cca0d47caec1f45739331dc354b6d749fdb8ab7c2b7f6cb24db39ca0c';
 
     const hashToSign = getAccountTransactionSignDigest(registerDataTransaction);
+
     const signature = Buffer.from(
         await ed.sign(hashToSign, wrongPrivateKey)
     ).toString('hex');