+++ packages/rust-bindings/src/external_functions.rs
@@ -31,6 +31,14 @@ pub fn get_credential_deployment_info_ext(signatures: &JsValue, unsigned_info: &
     }
 }
 
+#[wasm_bindgen(js_name = createEncryptedTransferData)]
+pub fn create_encrypted_transfer_ext(input: &str) -> String {
+    match create_encrypted_transfer_aux(input) {
+        Ok(s) => s,
+        Err(e) => format!("unable to create encrypted transfer due to: {}", e),
+    }
+}
+
 #[wasm_bindgen(js_name = deserializeState)]
 pub fn deserialize_state(contract_name: &str, state_bytes: String, schema: String) -> String {
     match deserialize_state_aux(contract_name, state_bytes, schema) {